requirements.txt
requests
python-dotenv
pillow

fastapi
uvicorn
jinja2
python-multipart

sqlalchemy

# Pour les tests - pas besoin pour l'image...
#pytest
#webdriver-manager
#selenium

#azure-ai-vision-imageanalysis